Moving distribution classes into a separate package
