+++ torch/distributions/bernoulli.py
@@ -1,0 +1,37 @@
+import torch
+from torch.distributions.distribution import Distribution
+
+
+class Bernoulli(Distribution):
+    r"""
+    Creates a Bernoulli distribution parameterized by `probs`.
+
+    Samples are binary (0 or 1). They take the value `1` with probability `p`
+    and `0` with probability `1 - p`.
+
+    Example::
+
+        >>> m = Bernoulli(torch.Tensor([0.3]))
+        >>> m.sample()  # 30% chance 1; 70% chance 0
+         0.0
+        [torch.FloatTensor of size 1]
+
+    Args:
+        probs (Tensor or Variable): the probabilty of sampling `1`
+    """
+
+    def __init__(self, probs):
+        self.probs = probs
+
+    def sample(self):
+        return torch.bernoulli(self.probs)
+
+    def sample_n(self, n):
+        return torch.bernoulli(self.probs.expand(n, *self.probs.size()))
+
+    def log_prob(self, value):
+        # compute the log probabilities for 0 and 1
+        log_pmf = (torch.stack([1 - self.probs, self.probs])).log()
+
+        # evaluate using the values
+        return log_pmf.gather(0, value.unsqueeze(0).long()).squeeze(0)
