+++ torch/distributions/utils.py
@@ -1,0 +1,13 @@
+from numbers import Number
+
+import torch
+
+
+def expand_n(v, n):
+    r"""
+    Cleanly expand float or Tensor or Variable parameters.
+    """
+    if isinstance(v, Number):
+        return torch.Tensor([v]).expand(n, 1)
+    else:
+        return v.expand(n, *v.size())
